Add Testcases
- Tests.py 추가
- nested routers 구현 (update requirements.txt)
- private subscription 제한

+++ apps/notice/views.py
@@ -7,77 +7,112 @@ from apps.channel.models import Channel
 from apps.notice.serializers import NoticeSerializer
 from apps.notice.permission import IsOwnerOrReadOnly
 
-
-
-class NoticeViewSet(viewsets.GenericViewSet):
+class NoticeIdViewSet(viewsets.GenericViewSet):
     queryset = Notice.objects.all()
     serializer_class = NoticeSerializer
-    permission_classes = (IsOwnerOrReadOnly, )
+    permission_classes = [
+        IsOwnerOrReadOnly()
+    ]
 
-    def create(self, request, pk):
+    def get_permissions(self):
+        return self.permission_classes
+
+    def create(self, request, channel_pk, pk=None):
         data = request.data.copy()
         data['writer'] = request.user.id
-        data['channel'] = Channel.objects.filter(id=pk)[0].id
+        channel = Channel.objects.filter(id=channel_pk).first()
 
+        if channel == None:
+            return Response({"error": "Wrong Channel ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        if not channel.managers.filter(id = request.user.id).exists():
+            return Response({"error": "Only managers can write a notice."}, status=status.HTTP_403_FORBIDDEN)
+        
+        data['channel'] = channel.id
         serializer = NoticeSerializer(data=data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-    
-    def list(self, request, pk):
+
+    def list(self, request, channel_pk):
         queryset = self.get_queryset()
         param = request.query_params
 
-        channel = Channel.objects.filter(id=pk)[0].id
-        data = queryset.select_related('channel')
+        channel = Channel.objects.filter(id = channel_pk).first()
+
+        if channel == None:
+            return Response({"error": "Wrong Channel ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        if channel.is_private and not (channel.managers.filter(id = request.user.id)):
+            return Response({"error": "This channel is private."}, status=status.HTTP_403_FORBIDDEN)
+
+        data = queryset.filter(channel = channel_pk)
+        recent_notices = 10
 
         if param.get('recent', '') == 'True':
-            if (data.count() > 10):
-                data = data[0:10]
+            if (data.count() > recent_notices):
+                data = data[0 : recent_notices]
 
         serializer = self.get_serializer(data, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-class NoticeIdViewSet(viewsets.GenericViewSet):
-    queryset = Notice.objects.all()
-    serializer_class = NoticeSerializer
-    permission_classes = (IsOwnerOrReadOnly, )
-
-    def list(self, request, pk, pk_2):
+    def retrieve(self, request, channel_pk, pk):
         queryset = self.get_queryset()
 
-        channel = Channel.objects.filter(id=pk)[0].id
-        data = queryset.select_related('channel')
+        channel = Channel.objects.filter(id = channel_pk).first()
 
-        notice = data.filter(id = pk_2)[0]
+        if channel == None:
+            return Response({"error": "Wrong Channel ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        if channel.is_private and not (channel.managers.filter(id = request.user.id)):
+            return Response({"error": "This channel is private."}, status=status.HTTP_403_FORBIDDEN)
+
+        data = queryset.filter(channel = channel_pk)
+        notice = data.filter(id = pk).first()
+
+        if notice == None:
+            return Response({"error": "Wrong Notice ID."}, status=status.HTTP_400_BAD_REQUEST)
 
         serializer = self.get_serializer(notice)
+        self.check_object_permissions(self.request, notice)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def patch(self, request, pk, pk_2):
+    def patch(self, request, channel_pk, pk):
         queryset = self.get_queryset()
 
-        channel = Channel.objects.filter(id=pk)[0].id
-        data = queryset.select_related('channel')
+        channel = Channel.objects.filter(id=channel_pk).first()
 
-        notice = data.filter(id = pk_2)[0]
+        if channel == None:
+            return Response({"error": "Wrong Channel ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        data = queryset.filter(channel = channel_pk)
+        notice = data.filter(id = pk).first()
+
+        if notice == None:
+            return Response({"error": "Wrong Notice ID."}, status=status.HTTP_400_BAD_REQUEST)
 
         data = request.data.copy()
 
         serializer = self.get_serializer(notice, data=data, partial=True)
         serializer.is_valid(raise_exception=True)
+        self.check_object_permissions(self.request, notice)
         serializer.update(notice, serializer.validated_data)
-        
         return Response(serializer.data, status=status.HTTP_200_OK)
     
-    @action(detail=True, methods=['DELETE'])
-    def delete(self, request, pk, pk_2):
+    def destroy(self, request, channel_pk, pk):
         queryset = self.get_queryset()
 
-        channel = Channel.objects.filter(id=pk)[0].id
-        data = queryset.select_related('channel')
+        channel = Channel.objects.filter(id=channel_pk).first()
 
-        notice = data.filter(id = pk_2)[0]
+        if channel == None:
+            return Response({"error": "Wrong Channel ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        data = queryset.filter(channel = channel_pk)
+        notice = data.filter(id = pk).first()
+
+        if notice == None:
+            return Response({"error": "Wrong Notice ID."}, status=status.HTTP_400_BAD_REQUEST)
+
+        self.check_object_permissions(self.request, notice)
         notice.delete()
-
         return Response(status = status.HTTP_200_OK)
